requirements.txt
django-ninja==0.13.2
websockets==9.1
uvicorn==0.14.0
psycopg2-binary==2.9.1
httpx==0.18.2
web3==5.20.1
ciso8601==2.1.3
pycoingecko==2.2.0
python-dotenv==0.18.0
channels==3.0.4
aiogram==2.14.2
tenacity~=8.0.1
PyJWT~=2.1.0
passlib[bcrypt]==1.7.4
celery==5.1.2
click==7.1.2
redis==3.5.3
django-celery-beat==2.2.1